+++ lxl-web/src/lib/components/ResourceImage.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import { type Image, Width } from '$lib/types/Auxd';
+	import { type Image, Width } from '$lib/types/auxd.types';
 	import placeholder from '$lib/assets/img/placeholder.svg';
 	import getTypeIcon from '$lib/utils/getTypeIcon';
 	import { bestSize } from '$lib/utils/auxd';
