+++ lxl-web/src/lib/components/find/SearchCard.svelte
@@ -1,14 +1,15 @@
 <script lang="ts">
 	import jmespath from 'jmespath';
-	import type { SearchResultItem } from '../(find)/search';
-	import DecoratedData from '$lib/components/DecoratedData.svelte';
-	import type { ResourceData } from '$lib/types/ResourceData';
-	import { ShowLabelsOptions } from '$lib/types/DecoratedData';
+	import type { SearchResultItem } from '$lib/types/search.types';
+	import type { ResourceData } from '$lib/types/resourceData.types';
+	import { LensType } from '$lib/types/xl.types';
+	import { ShowLabelsOptions } from '$lib/types/decoratedData.types';
+	import { LxlLens } from '$lib/types/display.types';
+
 	import { relativizeUrl } from '$lib/utils/http';
-	import { LensType } from '$lib/utils/xl';
-	import { LxlLens } from '$lib/utils/display.types';
-	import placeholder from '$lib/assets/img/placeholder.svg';
 	import getTypeIcon from '$lib/utils/getTypeIcon';
+	import placeholder from '$lib/assets/img/placeholder.svg';
+	import DecoratedData from '$lib/components/DecoratedData.svelte';
 	import { page } from '$app/stores';
 
 	export let item: SearchResultItem;
