+++ lxl-web/src/lib/components/find/SearchMapping.svelte
@@ -1,13 +1,15 @@
 <script lang="ts">
+	import { ShowLabelsOptions } from '$lib/types/decoratedData.types';
+	import type { DisplayMapping, SearchOperators } from '$lib/types/search.types';
+
 	import { page } from '$app/stores';
 	import DecoratedData from '$lib/components/DecoratedData.svelte';
-	import { ShowLabelsOptions } from '$lib/types/DecoratedData';
 	import { getModalContext } from '$lib/contexts/modal';
-	import type { DisplayMapping, SearchOperators } from './search';
 	import BiXLg from '~icons/bi/x-lg';
 	import BiPencil from '~icons/bi/pencil';
 	import BiPencilFill from '~icons/bi/pencil-fill';
 	import BiTrash from '~icons/bi/trash';
+
 	export let mapping: DisplayMapping[];
 	export let parentOperator: keyof typeof SearchOperators | undefined = undefined;
 	export let depth = 0;
