+++ lxl-web/src/lib/types/decoratedData.types.ts
