+++ lxl-web/src/lib/types/resourceData.types.ts
