+++ lxl-web/src/routes/(app)/[[lang=lang]]/[fnurgel=fnurgel]/+page.svelte
@@ -2,13 +2,14 @@
 	import { page } from '$app/stores';
 	import DecoratedData from '$lib/components/DecoratedData.svelte';
 	import InstancesList from './InstancesList.svelte';
-	import { ShowLabelsOptions } from '$lib/types/DecoratedData';
+	import { ShowLabelsOptions } from '$lib/types/decoratedData.types';
 	import getPageTitle from '$lib/utils/getPageTitle';
 	import Modal from '$lib/components/Modal.svelte';
 	import isFnurgel from '$lib/utils/isFnurgel';
 	import ResourceImage from '$lib/components/ResourceImage.svelte';
 	import { getHoldingsLink, handleClickHoldings } from './utils';
-	import { Width } from '$lib/utils/auxd.types';
+	import { Width } from '$lib/types/auxd.types';
+	import SearchResult from '$lib/components/find/SearchResult.svelte';
 
 	export let data;
 
@@ -227,6 +228,7 @@
 		</Modal>
 	{/if}
 </article>
+<SearchResult searchResult={$page.data.searchResult} />
 
 <style lang="postcss">
 	.resource {
